moved CalculPrix function
